Moved tests 2, 14, 30 and 52 into logictree_test.py
