+++ openquake/calculators/tests/classical_test.py
@@ -27,13 +27,13 @@ from openquake.hazardlib.sourcewriter import write_source_model
 from openquake.calculators.views import view, text_table
 from openquake.calculators.export import export
 from openquake.calculators.extract import extract
-from openquake.calculators.tests import CalculatorTestCase, NOT_DARWIN
+from openquake.calculators.tests import CalculatorTestCase
 from openquake.qa_tests_data.classical import (
-    case_01, case_02, case_12, case_14, case_18, case_19, case_22, case_23,
-    case_24, case_25, case_26, case_27, case_29, case_30, case_32, case_33,
+    case_01, case_12, case_18, case_19, case_22, case_23,
+    case_24, case_25, case_26, case_27, case_29, case_32, case_33,
     case_34, case_35, case_37, case_38, case_39, case_40, case_41,
     case_42, case_43, case_44, case_47, case_48, case_49,
-    case_50, case_51, case_52, case_53, case_54, case_55, case_57,
+    case_50, case_51, case_53, case_54, case_55, case_57,
     case_60, case_61, case_62, case_63, case_64, case_65,
     case_66, case_69, case_70, case_72, case_74, case_75, case_76, case_77,
     case_78, case_80, case_81, case_82, case_84)
@@ -109,16 +109,6 @@ class ClassicalTestCase(CalculatorTestCase):
             'SA(4.1) is out of the period range defined for [SadighEtAl1997]',
             str(ctx.exception))
 
-    def test_case_02(self):
-        self.run_calc(case_02.__file__, 'job.ini')
-
-        # check view inputs
-        lines = text_table(view('inputs', self.calc.datastore)).splitlines()
-        self.assertEqual(len(lines), 13)  # rst table with 13 rows
-
-        [fname] = export(('hcurves', 'csv'), self.calc.datastore)
-        self.assertEqualFiles('expected/hcurve.csv', fname)
-
     def test_case_12(self):
         # test Modified GMPE
         self.assert_curves_ok(
@@ -129,11 +119,6 @@ class ClassicalTestCase(CalculatorTestCase):
         df = self.calc.datastore.read_df('disagg_by_grp')
         fname = general.gettemp(text_table(df))
         self.assertEqualFiles('expected/disagg_by_grp.rst', fname)
-
-    def test_case_14(self):
-        # test classical with 2 gsims and 1 sample
-        self.assert_curves_ok(['hazard_curve-rlz-000_PGA.csv'],
-                              case_14.__file__)
 
     def test_case_18(self):  # GMPEtable, PointMSR, 3 hypodepths
         self.run_calc(case_18.__file__, 'job.ini',
@@ -257,20 +242,6 @@ class ClassicalTestCase(CalculatorTestCase):
 
         # then perform a classical calculation
         self.assert_curves_ok(['hazard_curve-PGA.csv'], case_29.__file__)
-
-    def test_case_30(self):
-        # point on the international data line
-        # this is also a test with IMT-dependent weights
-        if NOT_DARWIN:  # broken on macOS
-            self.assert_curves_ok(['hazard_curve-PGA.csv',
-                                   'hazard_curve-SA(1.0).csv'],
-                                  case_30.__file__)
-
-    def test_case_30_sampling(self):
-        # IMT-dependent weights with sampling by cheating
-        self.assert_curves_ok(
-            ['hcurve-PGA.csv', 'hcurve-SA(1.0).csv'],
-            case_30.__file__, number_of_logic_tree_samples='10', delta=1E-5)
 
     def test_case_32(self):
         # source specific logic tree
@@ -448,50 +419,6 @@ class ClassicalTestCase(CalculatorTestCase):
         self.assert_curves_ok(['hcurves-PGA.csv', 'hcurves-SA(0.2).csv',
                                'hcurves-SA(2.0).csv', 'uhs.csv'],
                               case_51.__file__)
-
-    def test_case_52(self):
-        # case with 2 GSIM realizations b1 (w=.9) and b2 (w=.1), 10 samples
-
-        # late_weights
-        self.run_calc(case_52.__file__, 'job.ini')
-        haz = self.calc.datastore['hcurves-stats'][0, 0, 0, 6]
-        aac(haz, 0.563831, rtol=1E-6)
-        ws = extract(self.calc.datastore, 'weights')
-        # sampled 8 times b1 and 2 times b2
-        aac(ws, [0.029412, 0.029412, 0.029412, 0.264706, 0.264706, 0.029412,
-                 0.029412, 0.264706, 0.029412, 0.029412], rtol=1E-5)
-
-        # early_weights
-        self.run_calc(case_52.__file__, 'job.ini',
-                      sampling_method='early_weights')
-        haz = self.calc.datastore['hcurves-stats'][0, 0, 0, 6]
-        aac(haz, 0.56355, rtol=1E-6)
-        ws = extract(self.calc.datastore, 'weights')
-        aac(ws, [0.1] * 10)  # all equal
-
-        # full enum, rlz-0: 0.554007, rlz-1: 0.601722
-        self.run_calc(case_52.__file__, 'job.ini',
-                      number_of_logic_tree_samples='0')
-        haz = self.calc.datastore['hcurves-stats'][0, 0, 0, 6]
-        aac(haz, 0.558779, rtol=1E-6)
-        ws = extract(self.calc.datastore, 'weights')
-        aac(ws, [0.9, 0.1])
-
-    def test_case_52_bis(self):
-        self.run_calc(case_52.__file__, 'job.ini',
-                      sampling_method='late_latin')
-        haz = self.calc.datastore['hcurves-stats'][0, 0, 0, 6]
-        aac(haz, 0.558779, rtol=1E-6)
-        ws = extract(self.calc.datastore, 'weights')
-        # sampled 5 times b1 and 5 times b2
-        aac(ws, [0.18, 0.02, 0.18, 0.18, 0.02, 0.02, 0.02, 0.02, 0.18, 0.18])
-
-        self.run_calc(case_52.__file__, 'job.ini',
-                      sampling_method='early_latin')
-        haz = self.calc.datastore['hcurves-stats'][0, 0, 0, 6]
-        aac(haz, 0.558779, rtol=1E-6)
-        ws = extract(self.calc.datastore, 'weights')
-        aac(ws, [0.1] * 10)  # equal weights
 
     def test_case_53(self):
         # Test case with 4-branch scaled backbone logic tree
