+++ openquake/qa_tests_data/logictree/case_02/expected_hazard_curves.xml
