+++ openquake/qa_tests_data/logictree/case_02/gsim_logic_tree.xml
