+++ openquake/qa_tests_data/logictree/case_02/lookup_sta.hdf5
