+++ openquake/qa_tests_data/logictree/case_14/AS2008_expected_curves.dat
