+++ openquake/qa_tests_data/logictree/case_14/CB2008_expected_curves.dat
