+++ openquake/qa_tests_data/logictree/case_30/ssm/shallow/gridded_seismicity_source_4.xml
