+++ openquake/qa_tests_data/logictree/case_30/ssm/shallow/int_kt.xml
