+++ openquake/qa_tests_data/logictree/case_30/ssm/shallow/shallow_faults.xml
