+++ openquake/qa_tests_data/logictree/case_30/ssmLT.xml
