+++ openquake/qa_tests_data/logictree/case_52/gsim_logic_tree.xml
