+++ openquake/qa_tests_data/logictree/case_52/job.ini
